CBBD-232: Fixed drug quantities to be decimals rather than integers

+++ bluebutton-data-pipeline-rif/src/main/java/gov/hhs/cms/bluebutton/datapipeline/rif/model/SNFClaimGroup.java
@@ -521,7 +521,7 @@ public final class SNFClaimGroup {
 		/**
 		 * @see Column#REV_CNTR_NDC_QTY
 		 */
-		public Optional<Integer> nationalDrugCodeQuantity;
+		public Optional<BigDecimal> nationalDrugCodeQuantity;
 
 		/**
 		 * @see Column#REV_CNTR_NDC_QTY_QLFR_CD
